requirements.txt
airtest==1.3.2
allure-pytest==2.13.2
beautifulsoup4==4.12.2
hytest==0.7.5
pocoui==1.0.94
pytest==7.4.3
Requests==2.31.0
pytest-assume==2.4.3
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-repeat==0.9.3
pytest-rerunfailures==13.0
